backend/requirements.txt
# Core
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9

# Database
sqlalchemy==2.0.30
alembic==1.13.1
aiosqlite==0.20.0

# Cache & Queue
redis==5.0.4
celery==5.4.0
flower==2.0.1

# Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.3

# Validation & Config
pydantic==2.7.1
pydantic-settings==2.2.1
email-validator==2.1.1

# AI & ML
openai==1.30.1
faiss-cpu==1.8.0
tiktoken==0.7.0
numpy==1.26.4

# Document Processing
PyPDF2==3.0.1
python-docx==1.1.2
chardet==5.2.0

# AWS
boto3==1.34.101
botocore==1.34.101

# HTTP
httpx==0.27.0
aiofiles==23.2.1

# Utilities
python-dotenv==1.0.1
structlog==24.1.0
tenacity==8.3.0
slowapi==0.1.9

# Testing
pytest==8.2.0
pytest-asyncio==0.23.6
httpx==0.27.0
